Correct bundler semantics
The install_sufia.sh script mimics the steps documented by the Sufia
project to install a vanilla Sufia 6.6.0 application.  As part of those
steps, a "rails new" action is performed to create the basis for the
new Sufia application.

A side-effect of "rails new" is that a Gemfile is created and a "bundle
install" performed to install the gems needed by Rails.  A side-effect
of that "bundle install" is that a Gemfile.lock is created, which fixes
the versions of certain gems used by Rails.

This update alters our subsequent use of "bundle install" (run after
adding the Sufia-related gems to Gemfile) to "bundle update".  This is
because the gem version requirements of the "sufia" gem may not be
compatible with those encapsulated in the aforementioned Gemfile.lock.
This was made apparent recently with the update of the "mime-types" gem
to v3, which caused the subsequent "bundle install" to fail.  Changing
this to "bundle update" is semantically more correct (as we have
altered the contents of Gemfile since the initial "bundle install" of
"rails new"), and allows bundler to re-evaluate all dependencies to
seek a suitable dependency resolution.

Finally, we also pin the version of the "active-fedora" gem to 9.9.1 as
it appears the latest 9.10 version of it causes runtime errors.
Pinning the gem to the earlier version allows Sufia to run.